dom/media/test/test_eme_stream_capture_blocked_case3.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test Encrypted Media Extensions</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
  <script type="text/javascript" src="eme.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
var manager = new MediaTestManager;

function startTest(test, token)
{
  // Case 3. capturing a media element that has MediaKeys should throw NotSupportedError.
  var p1 = new EMEPromise;
  var case3token = token + "_case3";
  let v3 = document.createElement("video");

  v3.addEventListener("loadeddata", function() {
    ok(true, TimeStamp(case3token) + " should reach loadeddata");

    for (const methodName of ["captureStream", "mozCaptureStream", "mozCaptureStreamUntilEnded"]) {
      let errorName = "";
      try {
        v3[methodName]();
      } catch (e) {
        errorName = e.name;
      }
      is(errorName, "NotSupportedError",
         TimeStamp(case3token) + ` ${methodName}() on EME element should throw NotSupportedError.`);
    }
    p1.resolve();
  });

  manager.started(case3token);
  let p2 = SetupEME(v3, test, case3token);

  Promise.all([p1.promise, p2])
  .catch(reason => ok(false, reason))
  .then(() => {
    CleanUpMedia(v3);
    manager.finished(case3token);
  });
}

SimpleTest.waitForExplicitFinish();
manager.runTests(gEMETests, startTest);
</script>
</pre>
</body>
</html>
